+++ extensions/microsoft-teams/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Microsoft Teams Changelog
 
+## [Update] - 2024-05-08
+
+# Features
+
+- Added ability to directly call with audio or video a user from a chat. 
+
 ## [Update] - 2023-05-16
 
 Fixes